requirements.txt
Pillow==6.0.0
Flask==1.0.3
